removed switch case and implemented String search

+++ Controller_Package/UserController.java
@@ -14,43 +14,37 @@ public class UserController {
     	userService.addUser("Aditya", "986723432", "aditya@gmail.com");
     	userService.addUser("Shorya", "9445432", "shorya@gmail.com");
     	userService.addUser("Shreya", "9445432456", "shreya@gmail.com");
-        int choice;
+        String choice;
         do {
             System.out.println("\n=== Contact Management System ===");
-            System.out.println("1. Add Contact");
-            System.out.println("2. View Contacts");
-            System.out.println("3. Search Contact");
-            System.out.println("4. Delete Contact");
-            System.out.println("5. Update Contact");
-            System.out.println("6. Exit");
+            System.out.println("Add Contact");
+            System.out.println("View Contacts");
+            System.out.println("Search Contact");
+            System.out.println("Delete Contact");
+            System.out.println("Update Contact");
+            System.out.println("Exit");
             System.out.print("Enter your choice: ");
 
-            choice = scanner.nextInt();
-            scanner.nextLine(); 
+            choice = scanner.nextLine();
+            choice = choice.toLowerCase();
 
-            switch (choice) {
-                case 1:
-                    addContact();
-                    break;
-                case 2:
-                    userService.viewUsers();
-                    break;
-                case 3:
-                    searchContact();
-                    break;
-                case 4:
-                    deleteContact();
-                    break;
-                case 5:
-                    updateContact();
-                    break;
-                case 6:
-                    System.out.println("Thank you for using Contact Manager. Goodbye!");
-                    break;
-                default:
-                    System.out.println("Invalid choice!");
+            if(choice.equals("add contact")) {
+                addContact();
+            } else if (choice.equals("view contacts")) {
+                userService.viewUsers();
+            } else if (choice.equals("search contact")) {
+                searchContact();
+            } else if (choice.equals("delete contact")) {
+                deleteContact();
+            } else if (choice.equals("update contact")) {
+                updateContact();
+            } else if (choice.equals("exit")) {
+                System.out.println("Thank you for using Contact Manager. Goodbye!");
+            } else {
+                System.out.println("Invalid choice! Please try again.");
             }
-        } while (choice != 6);
+
+        } while (!choice.equals("exit"));
     }
 
     private void addContact() {
